+++ packages/infra/src/api/routing.ts
@@ -7,8 +7,8 @@ TODO: uninteruptible commands! was for All except GET.
 */
 import type * as HttpApp from "@effect/platform/HttpApp"
 import { Rpc, RpcRouter } from "@effect/rpc"
-import type { NonEmptyArray } from "effect-app"
-import { Cause, Chunk, Context, Effect, FiberRef, flow, Layer, Predicate, S, Schema, Stream } from "effect-app"
+import type { NonEmptyArray, NonEmptyReadonlyArray } from "effect-app"
+import { Array, Cause, Chunk, Context, Effect, FiberRef, flow, Layer, Predicate, S, Schema, Stream } from "effect-app"
 import type { GetEffectContext, RPCContextMap } from "effect-app/client/req"
 import type { HttpServerError } from "effect-app/http"
 import { HttpRouter, HttpServerRequest, HttpServerResponse } from "effect-app/http"
@@ -216,14 +216,14 @@ export const makeMiddleware = <
   Context,
   CTXMap extends Record<string, RPCContextMap.Any>,
   RMW,
-  Layers extends Array<Layer.Layer.Any>
+  Layers extends NonEmptyReadonlyArray<Layer.Layer.Any> | never[]
 >(content: Middleware<Context, CTXMap, RMW, Layers>): Middleware<Context, CTXMap, RMW, Layers> => content
 
 export const makeRouter = <
   Context,
   CTXMap extends Record<string, RPCContextMap.Any>,
   RMW,
-  Layers extends Array<Layer.Layer.Any>
+  Layers extends NonEmptyReadonlyArray<Layer.Layer.Any> | never[]
 >(
   middleware: Middleware<Context, CTXMap, RMW, Layers>,
   devMode: boolean
@@ -287,14 +287,95 @@ export const makeRouter = <
 
     type Keys = keyof Filtered
 
-    const effect = <
+    type GetSuccess<Layers extends ReadonlyArray<Layer.Layer.Any>> = Layers extends
+      NonEmptyReadonlyArray<Layer.Layer.Any> ? {
+        [k in keyof Layers]: Layer.Layer.Success<Layers[k]>
+      }[number]
+      : never
+
+    type GetError<Layers extends ReadonlyArray<Layer.Layer.Any>> = Layers extends NonEmptyReadonlyArray<Layer.Layer.Any>
+      ? { [k in keyof Layers]: Layer.Layer.Error<Layers[k]> }[number]
+      : never
+
+    const effect: {
+      <
+        E,
+        R,
+        Make extends (
+          requests: RouteMatcher<CTXMap, Rsc, Context>
+        ) => Effect<
+          { [K in keyof Filter<Rsc>]: AHandler<Rsc[K]> },
+          any,
+          Strict extends true ? GetSuccess<TLayers> : any
+        >,
+        THandlers extends { [K in keyof Filter<Rsc>]: AHandler<Rsc[K]> },
+        TLayers extends NonEmptyReadonlyArray<Layer.Layer.Any> | never[],
+        Strict extends boolean = true
+      >(
+        layers: TLayers,
+        make: Make & ((requests: RouteMatcher<CTXMap, Rsc, Context>) => Effect<THandlers, E, R>),
+        strict?: Strict
+      ): {
+        moduleName: ModuleName
+        Router: HttpRouter.HttpRouter.TagClass<
+          RouterShape<Rsc>,
+          `${ModuleName}Router`,
+          never,
+          | Exclude<Context, HttpRouter.HttpRouter.Provided>
+          | Exclude<
+            RPCRouteR<
+              { [K in keyof Filter<Rsc>]: Rpc.Rpc<Rsc[K], _R<ReturnType<THandlers[K]["handler"]>>> }[keyof Filter<Rsc>]
+            >,
+            HttpRouter.HttpRouter.Provided
+          >
+        >
+        routes: Layer.Layer<
+          RouterShape<Rsc>,
+          Effect.Error<ReturnType<Make>> | GetError<TLayers>,
+          Exclude<R | RMW, GetSuccess<TLayers>>
+        >
+      }
+      <
+        E,
+        R,
+        THandlers extends { [K in keyof Filter<Rsc>]: AHandler<Rsc[K]> },
+        TLayers extends NonEmptyReadonlyArray<Layer.Layer.Any> | never[],
+        Strict extends boolean = true
+      >(
+        layers: [
+          ...TLayers,
+          Layer.Layer<Exclude<R, GetSuccess<TLayers>>, never, never>
+        ],
+        make: (requests: RouteMatcher<CTXMap, Rsc, Context>) => Effect<THandlers, E, R>,
+        strict?: Strict
+      ): {
+        moduleName: ModuleName
+        Router: HttpRouter.HttpRouter.TagClass<
+          RouterShape<Rsc>,
+          `${ModuleName}Router`,
+          never,
+          | Exclude<Context, HttpRouter.HttpRouter.Provided>
+          | Exclude<
+            RPCRouteR<
+              { [K in keyof Filter<Rsc>]: Rpc.Rpc<Rsc[K], _R<ReturnType<THandlers[K]["handler"]>>> }[keyof Filter<Rsc>]
+            >,
+            HttpRouter.HttpRouter.Provided
+          >
+        >
+        routes: Layer.Layer<
+          RouterShape<Rsc>,
+          E | GetError<TLayers>,
+          Exclude<R | RMW, GetSuccess<TLayers> | GetSuccess<Layers>>
+        >
+      }
+    } = (<
       E,
       R,
       THandlers extends {
         // import to keep them separate via | for type checking!!
         [K in Keys]: AHandler<Rsc[K]>
       },
-      TLayers extends NonEmptyArray<Layer.Layer.Any>
+      TLayers extends NonEmptyReadonlyArray<Layer.Layer.Any> | never[]
     >(
       layers: TLayers,
       make: (requests: typeof items) => Effect<THandlers, E, R>
@@ -429,7 +510,7 @@ export const makeRouter = <
       // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
       const routes = layer.pipe(
         Layer.provideMerge(r.Live),
-        layers ? Layer.provide(layers) as any : (_) => _,
+        layers && Array.isNonEmptyReadonlyArray(layers) ? Layer.provide(layers as any) as any : (_) => _,
         // TODO: only provide to the middleware?
         middleware.dependencies ? Layer.provide(middleware.dependencies as any) : (_) => _
       ) as Layer.Layer<
@@ -449,7 +530,7 @@ export const makeRouter = <
         Router: r,
         routes
       }
-    }
+    }) as any
 
     return effect
   }